requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
peewee==3.14.10
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
